# [Service] 구매 로또에 대한 결과값 받기
