+++ STEP_02/STUDY_NOTE.md
@@ -1,0 +1,8 @@
+## SETP2 STUDY_NOTY
+
+`SOLID 원칙`
+- SPR(단일 책임 원칙)
+- OCP(개방 폐쇄 원칙)
+- LSP(리스코프 치환 원칙)
+- DIP(의존 역전 원칙)
+- ISP(인터페이스 분리 원칙)
