+++ STEP_02/lotto/src/test/java/com/bootcamp/AppTest.java
@@ -1,8 +1,6 @@
 package com.bootcamp;
 
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.InputStream;
@@ -10,7 +8,6 @@ import java.io.OutputStream;
 import java.io.PrintStream;
 
 import com.bootcamp.lotto.service.InputService;
-import com.bootcamp.lotto.service.LottoService;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -20,12 +17,10 @@ import org.junit.jupiter.api.Test;
  */
 public class AppTest {
 
-    private static LottoService lottoService;
     private static InputService inputservice;
 
     @BeforeEach
     void initAll() {
-        lottoService = new LottoService();
         inputservice = new InputService();
     }
